fix lint and tests

+++ blockchain/chain/src/store/base_fee.rs
@@ -100,7 +100,7 @@ mod tests {
 
     use super::*;
 
-    fn construct_tests() -> Vec<(i64, i64, usize, i64, i64)> {
+    fn construct_tests() -> Vec<(i64, u64, usize, i64, i64)> {
         // (base_fee, limit_used, no_of_blocks, output)
         vec![
             (100_000_000, 0, 1, 87_500_000, 87_500_000),
@@ -115,14 +115,14 @@ mod tests {
             ),
             (
                 100_000_000,
-                BLOCK_GAS_LIMIT * 2,
+                BLOCK_GAS_LIMIT as u64 * 2,
                 2,
                 112_500_000,
                 112_500_000,
             ),
             (
                 100_000_000,
-                BLOCK_GAS_LIMIT * 15 / 10,
+                BLOCK_GAS_LIMIT as u64 * 15 / 10,
                 2,
                 110_937_500,
                 106_250_000,
